fix: refactored a little for code style and strucutre
