+++ components/atoms/progress/__tests__/stories.js
@@ -5,12 +5,11 @@
 // React
 import React from 'react'
 
+// Storybook
 import { boolean, number, select, text, withKnobs } from '@storybook/addon-knobs'
 
-import { CONTEXT } from '../../../'
-
 // UI
-import { Progress, ProgressBar } from '../'
+import { CONTEXT, Progress, ProgressBar } from '../../../'
 import Readme from '../README.md'
 
 export default {
@@ -28,15 +27,15 @@ const ProgressBarGroupId = 'ProgressBar'
 const ProgressBarGroupId2 = 'ProgressBar2'
 const ProgressGroupId = 'Progress'
 
-export const defaultStory = () => {
+export const main = () => {
   const now = number('now (percentage completed)', 30, {}, ProgressBarGroupId)
   return (
     <Progress height={text('height (with desired unit)', '1.5rem', ProgressGroupId)}>
       <ProgressBar
-        context={select('context', CONTEXT, 'primary', ProgressBarGroupId)}
-        striped={boolean('striped', false, ProgressBarGroupId)}
         animated={boolean('animated (striped must be checked)', false, ProgressBarGroupId)}
+        context={select('context', CONTEXT, 'primary', ProgressBarGroupId)}
         now={now}
+        striped={boolean('striped', false, ProgressBarGroupId)}
       >
         {now}%
       </ProgressBar>
@@ -44,19 +43,15 @@ export const defaultStory = () => {
   )
 }
 
-defaultStory.story = {
-  name: 'Default'
-}
-
 export const Striped = () => {
   const now = number('now (percentage completed)', 30, {}, ProgressBarGroupId)
   return (
     <Progress height={text('height (with desired unit)', '1.5rem', ProgressGroupId)}>
       <ProgressBar
-        context={select('context', CONTEXT, 'primary', ProgressBarGroupId)}
-        striped={boolean('striped', true, ProgressBarGroupId)}
         animated={boolean('animated (striped must be checked)', false, ProgressBarGroupId)}
+        context={select('context', CONTEXT, 'primary', ProgressBarGroupId)}
         now={now}
+        striped={boolean('striped', true, ProgressBarGroupId)}
       >
         {now}%
       </ProgressBar>
@@ -69,10 +64,10 @@ export const Animated = () => {
   return (
     <Progress height={text('height (with desired unit)', '1.5rem', ProgressGroupId)}>
       <ProgressBar
-        context={select('context', CONTEXT, 'primary', ProgressBarGroupId)}
-        striped={boolean('striped', true, ProgressBarGroupId)}
         animated={boolean('animated (striped must be checked)', true, ProgressBarGroupId)}
+        context={select('context', CONTEXT, 'primary', ProgressBarGroupId)}
         now={now}
+        striped={boolean('striped', true, ProgressBarGroupId)}
       >
         {now}%
       </ProgressBar>
@@ -86,18 +81,18 @@ export const MultipleProgressBars = () => {
   return (
     <Progress height={text('height (with desired unit)', '1.5rem', ProgressGroupId)}>
       <ProgressBar
-        context={select('context', CONTEXT, 'primary', ProgressBarGroupId)}
-        striped={boolean('striped', true, ProgressBarGroupId)}
         animated={boolean('animated (striped must be checked)', true, ProgressBarGroupId)}
+        context={select('context', CONTEXT, 'primary', ProgressBarGroupId)}
         now={now}
+        striped={boolean('striped', true, ProgressBarGroupId)}
       >
         {now}%
       </ProgressBar>
       <ProgressBar
-        context={select('context', CONTEXT, 'info', ProgressBarGroupId2)}
-        striped={boolean('striped', true, ProgressBarGroupId)}
         animated={boolean('animated (striped must be checked)', false, ProgressBarGroupId)}
+        context={select('context', CONTEXT, 'info', ProgressBarGroupId2)}
         now={now2}
+        striped={boolean('striped', true, ProgressBarGroupId)}
       >
         {now2}%
       </ProgressBar>
