+++ components/atoms/progress/components/ProgressBar.js
@@ -1,45 +1,61 @@
-import { number, string, bool } from 'prop-types'
-import styled, { css, keyframes } from 'styled-components'
-import { BACKGROUND } from '../../../'
+/**
+ * Progress Bar
+ */
 
-const ProgressBarStripesAnimation = keyframes`
-  0% {
-    background-position: 1rem 0;
-  }
-  100% {
-    background-position: 0 0;
-  }
-`
+// React
+import React from 'react'
+import { number, string, bool, oneOf } from 'prop-types'
 
-export const ProgressBarDiv = styled.div`
-  animation: ${({ animated }) =>
-    animated
-      ? css`
-          ${ProgressBarStripesAnimation} 1s linear infinite
-        `
-      : 'none 0 ease 0 1 normal none running'};
-  background-image: ${({ striped, bgImage }) => (striped ? bgImage : 'none')};
-  background-size: ${({ striped, bgSize }) => (striped ? bgSize : 'auto')};
-  color: ${({ color, context, theme }) =>
-    color || (context === 'white' ? theme.COLOUR.primary : theme.COLOUR.white)};
-  display: flex;
-  flex-direction: column;
-  width: ${({ now }) => now + '%'};
-  justify-content: center;
-  overflow: hidden;
-  text-align: center;
-  white-space: nowrap;
-  ${props => BACKGROUND(props)}
-  transition: ${({ transition }) => transition};
-`
+// UI
+import { CONTEXT, ProgressBarDiv } from '../../../'
 
-ProgressBarDiv.propTypes = {
+export const ProgressBar = props => {
+  return (
+    <ProgressBarDiv
+      aria-valuenow={props.now}
+      aria-valuemin={props.ariaValueMin}
+      aria-valuemax={props.ariaValueMax}
+      {...props}
+    />
+  )
+}
+
+ProgressBar.propTypes = {
   animated: bool,
+  ariaValueMax: number,
+  ariaValueMin: number,
   bgColor: string,
   bgImage: string,
   bgSize: string,
   color: string,
+  context: oneOf(Object.values(CONTEXT)),
   now: number,
+  role: string,
   striped: bool,
   transition: string
+}
+
+ProgressBar.defaultProps = {
+  animated: false,
+  ariaValueMax: 100,
+  ariaValueMin: 0,
+  bgColor: '#2D9CDB',
+  bgImage: `
+  linear-gradient(
+    45deg,
+    rgba(255,255,255,.15) 25%,
+    transparent 25%,
+    transparent 50%,
+    rgba(255,255,255,.15) 50%,
+    rgba(255,255,255,.15) 75%,
+    transparent 75%,
+    transparent
+    )
+    `,
+  bgSize: '1rem 1rem',
+  context: 'primary',
+  now: 0,
+  role: 'progressbar',
+  striped: false,
+  transition: 'width .6s ease'
 }
